Killed alterschema package, moved all contents to sqle/alterschema.go (ditto tests)
